requirements.txt
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
tqdm>=4.60.0
numpy>=1.20.0
tiktoken>=0.4.0
faiss-cpu>=1.7.0
llama-index>=0.8.0
openai>=1.0.0
google-generativeai>=0.3.0